refactored the undo button so that the last line within the lineArr is pushed into the redoArr and that element is spliced from linArr so that it does not appear when "drawing_changed " is called
added a redo button so that the most recent element that was added at the end of the redoArr is pushed to the lineArr so that it is displayed when "drawing-changed is" called. Then that line is spliced from the redoArr

did not realize that that the push method also incremenets the array size so I had to take out lineArr.length++; from the mouseDown event so that the length is accurate for the undo and button utilization.

(Step 4 Complete)